split Curve FieldElement ops
